I'm pretty pleased with RelaxedRoll (64 colors) and LazyRoll (256).
LazyRoll is new and based on a 256-color expansion of Roller, called BigRoller. BigRoller uses a few more hues on top of Roller, but has many more lightness levels, as well as extra desaturated colors. LazyRoll and RelaxedRoll both use Lloyd relaxation on the rigidly-defined Roller or BigRoller palette to fill in gaps.