Try lots of things to improve Wean/Wren.
I settled on increasing strength, which affects the blue noise and R2 effect, but then also halving the R2 effect's power. The error-diffusion w variables are weaker and don't incorporate populationBias as much. The limit for strength's effect is lower, and dMul is now a fixed 1.0/256.0 . This causes Wean not to preserve hue as well, but to have a less-flat appearance in some areas. The flat appearance was caused by some kind of repeating artifact, which could have been either error-diffusion or R2-based.