Adjust Blubber to work even with extreme palettes.
Even hyper-8, which has almost the maximum possible distance between colors on the RGB cube, can look pretty good with Blubber, now. It doesn't do this with Wean/Wren for an interesting reason: In Blubber, the current error is no longer included in the value that gets clamped, and is instead added afterwards. In Wean, error is included in the clamped value.

This also adds several new Lospec palettes. Prospecal-8 and Septembit23-6 are quite nice despite being very small palettes.