The new Snorgly palettes aren't as good!
I needed a way to show that using Oklab can be better than just RGB! And even snorgly uses Oklab a bit, for the comparisons of color similarity. There are also smaller single-purpose variants on AutomaticPalettizer; one runs a single or a few new palettes on all dithers, and one runs many palettes on one dither.